refactor: RequestURIFactory 분리
